refactor: util to shared helpers
