--- src/utils/index.ts
@@ -1,3 +1,0 @@
-import { type ActualDataListProps, getActualDataList } from './getActualDataList';
-
-export { type ActualDataListProps, getActualDataList };
